# packages/api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Public REST API of Honest Robin: Time. Authenticate with a personal access token (`Authorization: Bearer hrt_…`). Session-authenticated browser clients send the `HonestRobin-Account-Id` header to pick an account.",
    "license" : {
      "name" : "AGPL-3.0-only",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "title" : "Honest Robin: Time API",
    "version" : "v1"
  },
  "security" : [ {
    "token" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/auth/invite/accept" : {
      "post" : {
        "operationId" : "auth_accept_invite",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/invite/lookup" : {
      "post" : {
        "operationId" : "auth_lookup_invite",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InviteInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "auth_login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
...
    "/api/v1/auth/magic_link" : {
      "post" : {
        "operationId" : "auth_magic_link",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/magic_link/consume" : {
      "post" : {
        "operationId" : "auth_consume_magic_link",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/password_reset" : {
      "post" : {
        "operationId" : "auth_password_reset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/password_reset/consume" : {
      "post" : {
        "operationId" : "auth_consume_password_reset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/reauth" : {
      "post" : {
        "operationId" : "auth_reauth",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReauthRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/signup" : {
      "post" : {
        "operationId" : "auth_signup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/two_factor" : {
      "post" : {
        "operationId" : "auth_two_factor",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SecondFactorRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SignedInResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
    "/api/v1/auth/verify_email" : {
      "post" : {
        "operationId" : "auth_verify_email",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "tags" : [ "auth" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AcceptInviteRequest" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "token" ]
      },
...
      "EmailRequest" : {
        "properties" : {
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "email" ]
      },
...
      "InviteInfo" : {
        "properties" : {
          "account_name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "user_exists" : {
            "type" : "boolean"
          }
        },
        "required" : [ "account_name", "email", "name", "user_exists" ]
      },
...
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ]
      },
...
      "PasswordResetRequest" : {
        "properties" : {
          "password" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "password", "token" ]
      },
...
      "ReauthRequest" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },
...
      "SecondFactorRequest" : {
        "properties" : {
          "challenge" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "recovery_code" : {
            "type" : "string"
          }
        },
        "required" : [ "challenge" ]
      },
...
      "SignedInResponse" : {
        "properties" : {
          "account_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "challenge" : {
            "type" : "string"
          },
          "two_factor_required" : {
            "type" : "boolean"
          },
          "user_id" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "two_factor_required", "user_id" ]
      },
      "SignupRequest" : {
        "properties" : {
          "account_name" : {
            "type" : "string"
          },
          "default_currency" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          },
          "week_start" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "account_name", "email", "name", "password" ]
      },
...
      "TokenRequest" : {
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "token" ]
      },
...
    },
    "securitySchemes" : {
      "token" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}